Replace add-flow command with set-flows command
This command reads flows in JSON from a file and passes
them to the kernel. This allows setting of all flow attributes
supported by libflow-table's JSON support: all attributes in the alpha
Flow Table API at the time of writing.

This is in contrast to the add-flow command which is removed which
only allowed setting of a very limited set of attributes